Content/Python/requirements.txt
# Yes UE FSD - Two-Layer Test Framework Requirements
# Python 3.7+

# Testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Process management
psutil>=5.9.0

# Optional: Enhanced reporting
pytest-html>=3.1.0  # HTML test reports
pytest-json-report>=1.5.0  # JSON test reports
allure-pytest>=2.13.0  # Allure test reports with rich visualizations
pytest-metadata>=3.0.0  # Add metadata to test reports

# Optional: Coverage
pytest-cov>=4.0.0  # Code coverage

# Optional: Parallel execution
pytest-xdist>=3.0.0  # Run tests in parallel

# BDD support
pytest-bdd>=6.1.0  # Gherkin-style BDD testing

# Test retry and flake detection
pytest-rerunfailures>=12.0  # Retry failed tests automatically

# Visual regression testing
Pillow>=10.0.0  # Image processing

# Data-driven testing
pyyaml>=6.0  # YAML data loading

# API testing
requests>=2.31.0  # HTTP client library
